requirements.txt
torchvision~=0.2.1
pytorch~=1.9.0.dev20210529
numpy~=1.19.2
matplotlib~=3.2.2
scikit-learn~=0.23.2